Split core/Gradient.h contents into new files
