+++ drake/CMakeLists.txt
@@ -271,7 +271,6 @@ include_directories(thirdParty/bsd/spruce/include)
 
 add_subdirectory(common)
 add_subdirectory(math)
-add_subdirectory(core)
 add_subdirectory(util)
 add_subdirectory(solvers)
 add_subdirectory(systems)
