--- drake/core/CMakeLists.txt
@@ -1,1 +1,0 @@
-drake_install_headers(Gradient.h)
