+++ drake/math/test/expmap_test.cc
@@ -3,12 +3,11 @@
 #include <unsupported/Eigen/AutoDiff>
 #include "gtest/gtest.h"
 
-#include "drake/core/Gradient.h"
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
 #include "drake/util/eigen_matrix_compare.h"
 
-using drake::initializeAutoDiff;
+using drake::math::initializeAutoDiff;
 using Eigen::AutoDiffScalar;
 using Eigen::Matrix3d;
 using Eigen::Vector4d;
