+++ drake/util/quat2expmapmex.cpp
@@ -2,15 +2,17 @@
 
 #include "drake/math/autodiff_gradient.h"
 #include "drake/math/expmap.h"
+#include "drake/math/gradient.h"
 #include "drake/util/mexify.h"
 #include "drake/util/standardMexConversions.h"
 #include "drake/util/drakeGeometryUtil.h"
 #include "drake/util/makeFunction.h"
-#include "drake/core/Gradient.h"
 
 using namespace std;
 using namespace Eigen;
 using namespace drake;
+
+using drake::math::initializeAutoDiff;
 
 pair<Vector3d, typename Gradient<Vector3d, 4>::type> quat2expmapWithGradient(
     const MatrixBase<Map<const Vector4d>>& quat) {
