+++ drake/util/unwrapExpmapmex.cpp
@@ -2,11 +2,11 @@
 
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
+#include "drake/math/gradient.h"
 #include "drake/util/mexify.h"
 #include "drake/util/standardMexConversions.h"
 #include "drake/util/drakeGeometryUtil.h"
 #include "drake/util/makeFunction.h"
-#include "drake/core/Gradient.h"
 
 using namespace std;
 using namespace Eigen;
@@ -14,6 +14,7 @@ using namespace drake;
 
 using drake::math::autoDiffToValueMatrix;
 using drake::math::autoDiffToGradientMatrix;
+using drake::math::initializeAutoDiff;
 
 // note: gradient only w.r.t. expmap2...
 pair<Vector3d, typename Gradient<Vector3d, 3>::type> unwrapExpmapWithGradient(
